backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.20.0
httpx==0.26.0
anthropic==0.17.0
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
python-whois==0.9.4
beautifulsoup4==4.12.3
lxml==5.1.0
googlemaps==4.10.0
pandas==2.2.0
openpyxl==3.1.2
websockets==12.0
